Move integer encoding/decoding tests to new file.
