+++ test/test_hpack.py
@@ -1,8 +1,5 @@
 # -*- coding: utf-8 -*-
-from hpack.hpack import (
-    Encoder, Decoder, encode_integer, decode_integer, _dict_to_iterable,
-    _to_bytes
-)
+from hpack.hpack import Encoder, Decoder, _dict_to_iterable, _to_bytes
 from hpack.exceptions import (
     HPACKDecodingError, InvalidTableIndex, OversizedHeaderListError
 )
@@ -639,47 +636,6 @@ class TestHPACKDecoder(object):
 
         with pytest.raises(OversizedHeaderListError):
             d.decode(data)
-
-
-class TestIntegerEncoding(object):
-    # These tests are stolen from the HPACK spec.
-    def test_encoding_10_with_5_bit_prefix(self):
-        val = encode_integer(10, 5)
-        assert len(val) == 1
-        assert val == bytearray(b'\x0a')
-
-    def test_encoding_1337_with_5_bit_prefix(self):
-        val = encode_integer(1337, 5)
-        assert len(val) == 3
-        assert val == bytearray(b'\x1f\x9a\x0a')
-
-    def test_encoding_42_with_8_bit_prefix(self):
-        val = encode_integer(42, 8)
-        assert len(val) == 1
-        assert val == bytearray(b'\x2a')
-
-
-class TestIntegerDecoding(object):
-    # These tests are stolen from the HPACK spec.
-    def test_decoding_10_with_5_bit_prefix(self):
-        val = decode_integer(b'\x0a', 5)
-        assert val == (10, 1)
-
-    def test_encoding_1337_with_5_bit_prefix(self):
-        val = decode_integer(b'\x1f\x9a\x0a', 5)
-        assert val == (1337, 3)
-
-    def test_encoding_42_with_8_bit_prefix(self):
-        val = decode_integer(b'\x2a', 8)
-        assert val == (42, 1)
-
-    def test_decode_empty_string_fails(self):
-        with pytest.raises(HPACKDecodingError):
-            decode_integer(b'', 8)
-
-    def test_decode_insufficient_data_fails(self):
-        with pytest.raises(HPACKDecodingError):
-            decode_integer(b'\x1f', 5)
 
 
 class TestDictToIterable(object):
